requirements.txt
crewai>=0.19.0
python-dotenv>=1.0.0
litellm>=1.26.1
pandas~=2.2.3
numpy~=1.26.4
pydantic~=2.9.2
chromadb~=0.4.24
sqlalchemy~=2.0.35
scipy~=1.14.1
statsmodels~=0.14.4
rich~=13.9.4
pytest~=8.3.3